Massive refactor of the `repex.py` file, which no longer exists
* Split out the `repex.py` into its own `sampling` subpackage
    * `MultiStateSampler`
    * `MultiStateReporter`
    * `ReplicaExchangeSampler`
    * `ParallelTemperingSampler`
* Divided the classes of `repex.py` into their own modules
* Decoupled most of `sampling` from yank-specific things for eventual migration to OpenMMTools
    * Hooks into MPI module
    * Hooks into a logger config from utils
    * Moved the `SimulationNaNError` to `samplingutils`
    * Custom YAML Loader/Dumpers contained in `sampling`
    * Moved `mixing` to submodule of `sampling`
 * Tried to fix all the tests I could
 * Most references to classes in `sampling` should be `sampling.{class}`
 * Some other PEP8 cleanup as I found it

 Things which still need done:
 * Update and verify all Doc pages
 * Actually run and ensure tests work
