+++ Yank/mpi.py
@@ -60,7 +60,6 @@ logger = logging.getLogger(__name__)
 disable_mpi = False
 
 
-
 # ==============================================================================
 # MAIN FUNCTIONS
 # ==============================================================================
