+++ Yank/tests/test_experiment.py
@@ -13,22 +13,22 @@ Test YAML functions.
 # GLOBAL IMPORTS
 # =============================================================================================
 
-import time
-import shutil
-import textwrap
-import unittest
-import tempfile
 import itertools
+import shutil
+import tempfile
+import textwrap
+import time
+import unittest
 
 import mdtraj
 import numpy as np
-
-from nose.tools import assert_raises, assert_equal, assert_raises_regexp
 from nose.plugins.attrib import attr
+from nose.tools import assert_raises, assert_equal, assert_raises_regexp
 
 from yank.experiment import *
+
 # silence the citations at a global level
-repex.ReplicaExchange._global_citation_silence = True
+sampling.MultiStateSampler._global_citation_silence = True
 
 
 # ==============================================================================
@@ -2364,7 +2364,7 @@ def solvation_stock(tmp_dir, overwrite_options=None):
     assert os.path.isdir(output_dir)
     for solvent in ['solvent1.nc', 'solvent2.nc']:
         solvent_path = os.path.join(output_dir, solvent)
-        reporter = repex.Reporter(solvent_path, open_mode=None)
+        reporter = sampling.MultiStateReporter(solvent_path, open_mode=None)
         assert reporter.storage_exists()
         del reporter
     assert os.path.isfile(os.path.join(output_dir, 'experiments.yaml'))
