+++ Yank/utils.py
@@ -1827,15 +1827,6 @@ class TLeap:
 
 
 # =============================================================================================
-# YANK Exceptions
-# =============================================================================================
-
-class SimulationNaNError(Exception):
-    """Error when a simulation goes to NaN"""
-    pass
-
-
-# =============================================================================================
 # Main and tests
 # =============================================================================================
