+++ docs/api/index.rst
@@ -14,8 +14,8 @@ YAML input files, please see the command line interface and the YAML options.
    experiment_api
    mpi_api
    pipeline_api
-   repex_api
    restraints_api
+   sampling_api
    utils_api
    yank_api
