--- docs/api/repex_api.rst
@@ -1,5 +1,0 @@
-.. _API_repex:
-
-.. automodule:: yank.repex
-   :members:
-   :inherited-members:
