+++ docs/api/sampling_api.rst
@@ -1,0 +1,5 @@
+.. _API_sampling:
+
+.. automodule:: yank.sampling
+   :members:
+   :inherited-members:
